extract stubs to the test utils
