+++ metrics/web/test/features/dashboard/presentation/page/dashboard_page_test.dart
@@ -13,6 +13,8 @@ import 'package:metrics/features/dashboard/presentation/widgets/circle_percentag
 import 'package:metrics_core/metrics_core.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
 
+import '../../../../test_utils/metrics_store_stub.dart';
+
 void main() {
   group("DashboardPage", () {
     testWidgets(
@@ -150,35 +152,6 @@ class DashboardTestbed extends StatelessWidget {
       ),
     );
   }
-}
-
-class MetricsStoreStub implements ProjectMetricsStore {
-  static const _testProjectMetrics = ProjectMetricsData(
-    projectId: '1',
-    projectName: 'project',
-    coverage: Percent(0.4),
-    stability: Percent(0.7),
-    buildNumberMetric: 1,
-    averageBuildDurationInMinutes: 1,
-    performanceMetrics: [],
-    buildResultMetrics: [],
-  );
-
-  final List<ProjectMetricsData> projectMetrics;
-
-  const MetricsStoreStub({
-    this.projectMetrics = const [_testProjectMetrics],
-  });
-
-  @override
-  Stream<List<ProjectMetricsData>> get projectsMetrics =>
-      Stream.value(projectMetrics);
-
-  @override
-  Future<void> subscribeToProjects() async {}
-
-  @override
-  void dispose() {}
 }
 
 class MetricsStoreErrorStub extends MetricsStoreStub {
